일단 자바스크립트에서 타입스크립트로 변환 가능한 코드는 변환함... 다른 코드도 허락해준다면 변환할 예정
